% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleMB_analysisModule.R
\name{calcAlpha}
\alias{calcAlpha}
\title{calculate Alpha diversity}
\usage{
calcAlpha(
  seurat.object,
  measures = c("Simpson", "Shannon"),
  tax_level = "species"
)
}
\arguments{
\item{seurat.object}{seurat object}

\item{measures}{Default :\code{c("Simpson", "Shannon")}.
Meaning that all available alpha-diversity measures will be included.
Could selected form \code{c("Observed", "Chao1", "ACE", "Shannon", "Simpson", "InvSimpson", "Fisher")}.
Details see \code{\link[phyloseq]{estimate_richness}} from package \code{\link{phyloseq}}.}

\item{tax_level}{Default: \code{"species"} taxonomy level.}
}
\value{
seurat object
}
\description{
calculate Alpha diversity
}
\examples{
data(seurat.object)
seurat.object <- calcAlpha(seurat.object, tax_level = "genus")
head(seurat.object@meta.data)

}
