% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleMB_analysisModule.R
\name{calcCorr}
\alias{calcCorr}
\title{Title}
\usage{
calcCorr(
  seurat.object,
  host.feature = "marker",
  microbe.feature = "all",
  group.by = "seurat_clusters"
)
}
\arguments{
\item{seurat.object}{seurat.object}

\item{host.feature}{Default: \code{"marker_gene"}
\describe{
\item{\code{"marker_gene"}}{Using maker genes of seurat_clusters.}
}}

\item{microbe.feature}{Default: \code{"all"}
\describe{
\item{\code{"all"}}{Using all species.}
}}

\item{group.by}{Default: \code{"seurat_clusters"}}
}
\value{
seurat object with correlation analysis result in \code{seurat.object@misc$corr$corr_result} solt.
}
\description{
Title
}
\examples{
NULL
}
