% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleMB_plotModule.R
\name{plotCorr}
\alias{plotCorr}
\title{Title}
\usage{
plotCorr(seurat.object, microbe_list = 1:10, feature_list = 1:20)
}
\arguments{
\item{seurat.object}{seurat object}

\item{microbe_list}{Default: \code{1:10}.}

\item{feature_list}{Default: \code{1:20}.}
}
\value{
seurat object
}
\description{
Title
}
\examples{
NULL

}
