% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleMB_readData.R
\name{read_Host}
\alias{read_Host}
\title{Title}
\usage{
read_Host(named.outs.dirs)
}
\arguments{
\item{named.outs.dirs}{Default: String or string vector, representing the paths to the outs directories of the output results from Cell Ranger or Spacer Ranger.}
}
\value{
seurat object
}
\description{
Title
}
\examples{
named.outs.dirs <- c(
  system.file("extdata",
    "sc/SAMN20336351/outs/",
    package = "SingleMBAnalyst"
  ),
  system.file("extdata",
    "sc/SAMN20336352/outs/",
    package = "SingleMBAnalyst"
  )
)
names(named.outs.dirs) <- c("SAMN20336351", "SAMN20336352")
seurat.object <- read_Host(named.outs.dirs)
}
