% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleMB_readData.R
\name{read_Microbe}
\alias{read_Microbe}
\title{Title}
\usage{
read_Microbe(named.outs.dirs, align.method = "kraken2")
}
\arguments{
\item{named.outs.dirs}{String or string vector, representing the paths to the outs directories of the output results from Cell Ranger or Spacer Ranger.}

\item{align.method}{Default: \code{"kraken2"}. SingleMB alignment methods, selected from ["kraken2", "blast"]}
}
\value{
microbe object
}
\description{
Title
}
\examples{
named.outs.dirs <- c(
  system.file("extdata",
    "sc/SAMN20336351/outs/",
    package = "SingleMBAnalyst"
  ),
  system.file("extdata",
    "sc/SAMN20336352/outs/",
    package = "SingleMBAnalyst"
  )
)
names(named.outs.dirs) <- c("SAMN20336351", "SAMN20336352")
microbe.phyloseq.object <- read_Microbe(named.outs.dirs)
}
