% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleMB_readData.R
\name{read_singleMB}
\alias{read_singleMB}
\title{Title}
\usage{
read_singleMB(named.outs.dirs, align.method = "kraken2", max.dim = 30000)
}
\arguments{
\item{named.outs.dirs}{String or string vector, representing the paths to the outs directories of the output results from Cell Ranger or Spacer Ranger.}

\item{align.method}{Default: \code{"kraken2"}. SingleMB alignment methods, selected from ["kraken2", "blast"]}

\item{max.dim}{Default: \code{30000}.
Return phyloseq object if cell numbers <= \strong{\code{max.dim}}, else return seurat object.}
}
\value{
seurat object with microbe data. "seurat.object@misc$microbe"
}
\description{
Title
}
\examples{
named.outs.dirs <- c(
  system.file("extdata",
    "sc/SAMN20336351/outs/",
    package = "SingleMBAnalyst"
  ),
  system.file("extdata",
    "sc/SAMN20336352/outs/",
    package = "SingleMBAnalyst"
  )
)
names(named.outs.dirs) <- c("SAMN20336351", "SAMN20336352")
seurat.object <- read_singleMB(named.outs.dirs)

}
