% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleMB_analysisModule.R
\name{runLEfSe}
\alias{runLEfSe}
\title{Title}
\usage{
runLEfSe(seurat.object, group.by = "seurat_clusters", compairs = c("0", "3"))
}
\arguments{
\item{seurat.object}{seurat.object}

\item{group.by}{Default: \code{"seurat_clusters"}.}

\item{compairs}{Default: \code{c("0", "3")}}
}
\value{
seurat object with LEfSe analysis result in \code{seurat.object@misc$microbe$lefse} solt.
}
\description{
Title
}
\examples{
NULL
}
